refactor: rename variant.py to rope_variant.py
